complete

+++ components/ui/header.tsx
@@ -5,9 +5,12 @@ import Link from 'next/link';
 import Logo from '@/components/ui/logo';
 import { jwtDecode } from "jwt-decode";
 import { JwtPayload } from "../types";
+import { BACKEND_URL } from '../constant';
+import axios, { AxiosError } from 'axios';
 
 export default function Header() {
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [isAuthenticated, setIsAuthenticated] = useState<boolean>(false);
+  const [loading, setLoading] = useState<boolean>(false);
 
   useEffect(() => {
     const token = localStorage.getItem('authToken');
@@ -28,10 +31,31 @@ export default function Header() {
     }
   }, [isAuthenticated]);
 
-  const handleSignOut = () => {
-    localStorage.clear();
-    setIsAuthenticated(false);
-    window.location.href = "/";
+  const SignOut = async (userEmail: string) => {
+    try {
+      setLoading(true);
+      const result = await axios.get(`${BACKEND_URL}/auth/signout?email=${encodeURIComponent(userEmail)}`);
+      console.log(result);
+    } catch (err) {
+      console.error(err as AxiosError)
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  const handleSignOut = async () => {
+    const token = localStorage.getItem('authToken');
+    if (token) {
+      const decodedToken = jwtDecode<JwtPayload>(token);
+      if (decodedToken.email !== "") {
+        localStorage.clear();
+        setIsAuthenticated(false);
+        await SignOut(decodedToken.email)
+        window.location.href = "/";
+      } else {
+        console.error("No user found")
+      }
+    }
   };
 
   return (
@@ -71,7 +95,7 @@ export default function Header() {
                   onClick={handleSignOut}
                   className="cursor-pointer text-sm font-medium text-green-100 px-3 lg:px-5 py-2 flex items-center"
                 >
-                  Sign out
+                  {loading ? "Signing Out..." : "Sign Out"}
                 </button>
               </li>
             ) : (
